feat(workloads): refactor identities into own TF module
Previously the AKS module leveraged control plane identities. Now with
workload identities, we'll need to create per "workspace" or "app"
- namespace
- service account for workload
- managed identity for the workload
- federated credential that creates trust relationship between the k8s
  managed service account and the ARM managed managed identity.

Benefit is that this is cleaner separation security wise. BUT
disadvantage is also that on the resource side, e.g. key vault,
container registry, it's still shared resources. So the role
assignments are in effect duplicates.